+++ Localization/deDE.lua
@@ -124,17 +124,20 @@ L.Settings.FeatureFlagLabels = {
 	[Private.Enum.FeatureFlag.UseTargetClassColor] = "Färbe Namen in Klassenfarbe ein",
 	[Private.Enum.FeatureFlag.ShowInterruptSource] = "Zeige unterbrechenden Spielernamen",
 	[Private.Enum.FeatureFlag.UseInterruptSourceClassColor] = "Färbe Namen in Klassenfarbe ein",
-	[Private.Enum.FeatureFlag.PlayFocusTTSReminder] = "Text-zu-Sprache Erinnerung abspielen",
+	[Private.Enum.FeatureFlag.PlayTargetingTTSReminder] = "Text-zu-Sprache Erinnerung abspielen",
 	[Private.Enum.FeatureFlag.IgnoreFriendlies] = "Freundliche Ziele ignorieren",
 	[Private.Enum.FeatureFlag.UnfillChannels] = "Kanalisierungszauberanimation leert statt füllt",
-	[Private.Enum.FeatureFlag.HideWhenUninterruptible] = "Nur anzeigen wenn Zauber unterbrechbar & kann unterbrechen",
+	[Private.Enum.FeatureFlag.HideWhenUninterruptible] = "Nur anzeigen wenn unterbrechbar & kann unterbrechen",
+	[Private.Enum.FeatureFlag.PlaySoundOnCastStart] = "Text-zu-Sprache bei Beginn eines Zaubers abspielen",
 }
 L.Settings.FeatureFlagSettingTitles = {
 	[Private.Enum.FeatureFlag.ShowIcon] = "'Features' links hat weitere Infos im Tooltip",
 	[Private.Enum.FeatureFlag.ShowTargetName] = "Anvisiertes Ziel",
 	[Private.Enum.FeatureFlag.ShowInterruptSource] = "Unterbrechungsquelle",
-	[Private.Enum.FeatureFlag.UnfillChannels] = "Diverses",
+	[Private.Enum.FeatureFlag.PlaySoundOnCastStart] = "Toneinstellungen",
+	[Private.Enum.FeatureFlag.UnfillChannels] = "Weiteres",
 }
+L.Settings.CastStartText = "Wirkt"
 
 do
 	local function CreateFeatureFlagTooltip(id, tooltip)
@@ -148,8 +151,15 @@ do
 		),
 		CreateFeatureFlagTooltip(Private.Enum.FeatureFlag.UseTargetClassColor, "Kann Kontrastprobleme haben."),
 		CreateFeatureFlagTooltip(
-			Private.Enum.FeatureFlag.PlayFocusTTSReminder,
+			Private.Enum.FeatureFlag.PlayTargetingTTSReminder,
 			"Spielt 'Fokus' (oder 'Ziel') Text-zu-Sprache ab wenn die derzeitg beobachtete Einheit verschwindet, bspw. weil sie getötet wurde, zur Erinnerung ein neues Ziel zu wählen. Nur in Instanzen aktiv!"
+		),
+		CreateFeatureFlagTooltip(
+			Private.Enum.FeatureFlag.PlaySoundOnCastStart,
+			string.format(
+				"Spielt '%s' Text-zu-Sprache ab wenn die beobachtete Einheit anfängt zu wirken. Bedingungslos, kann nicht prüfen ob du unterbrechen kannst oder ob der Zauber unterbrechbar ist. Nur in Instanzen!",
+				L.Settings.CastStartText
+			)
 		),
 		CreateFeatureFlagTooltip(
 			Private.Enum.FeatureFlag.UnfillChannels,