+++ Types.lua
@@ -140,6 +140,7 @@
 ---@field SetTargetNameVisibility fun(self: AdvancedFocusCastBarMixin, bool: boolean)
 ---@field GetMaybeColoredUnitName fun(self: AdvancedFocusCastBarMixin, unit: string): string
 ---@field SetAlphaFromFeatureFlag fun(self: AdvancedFocusCastBarMixin, interruptDuration: LuaDurationObject)
+---@field PlayTTS fun(self: AdvancedFocusCastBarMixin, text: string)
 
 -------- library types